finish simple ORM

+++ mySimpleORM/src/test/java/org/bruno/mySimpleORM/House.java
@@ -2,14 +2,14 @@ package org.bruno.mySimpleORM;
 
 public class House {
 
-    private int kvartir;
-    private int podezdov;
+    private long kvartir;
+    private long podezdov;
     private String xozyain;
     private String[] blijaiwieulitci;
     private String ulitca;
-    private int[] neplatelwiki;
+    private long[] neplatelwiki;
 
-    public House(int kvartir, int podezdov, String xozyain, String[] blijaiwieulitci, String ulitca, int[] neplatelwiki) {
+    public House(long kvartir, long podezdov, String xozyain, String[] blijaiwieulitci, String ulitca, long[] neplatelwiki) {
         this.kvartir = kvartir;
         this.podezdov = podezdov;
         this.xozyain = xozyain;
